refactor(profile): move `@top/@artists` to `views`
