+++ app/profile/[id]/views/index.ts
@@ -1,1 +1,2 @@
 export * from './top-genres-view'
+export * from './top-artists-view'
